Send OC Go Span data in App Insights format to 1Agent via json socket
